Excel Sheets added to download page for admin

+++ src/contexts/userContext.tsx
@@ -1,42 +1,57 @@
-import React, { ReactNode, createContext, useContext, useState } from "react";
+import React, { ReactNode, createContext, useContext, useState, useEffect } from "react";
 import Axios from "../Axios";
 
 interface User {
   name: string;
   email: string;
-  role:string;
-  profileImage:string
+  role: string;
+  profileImage: string;
 }
 
 interface AuthContextType {
   user: User | null;
   setUser: React.Dispatch<React.SetStateAction<User | null>>;
-  checkUserLoggedIn: any;
+  checkUserLoggedIn: () => Promise<void>;
+  loading: boolean;
 }
 
 const AuthContext = createContext<AuthContextType | null>(null);
 
 const AuthProvider = ({ children }: { children: ReactNode }) => {
   const [user, setUser] = useState<User | null>(null);
+  const [loading, setLoading] = useState<boolean>(true); // Added loading state
 
   const checkUserLoggedIn = async () => {
     const token = localStorage.getItem("token");
+    if (!token) {
+      setLoading(false); // No token, no need to call the server
+      return;
+    }
+
     try {
-      // Send token to server to check login status
+      setLoading(true); // Start loading
       const response = await Axios.post("/teacher/check-login", { token });
       setUser(response.data.user);
+
       if (!response.data.loggedIn) {
         localStorage.clear();
-        setUser(response.data.user);
-        window.location.href = "/login";
+        window.location.href = "/login"; // Redirect if not logged in
       }
     } catch (error: any) {
       console.log(error.response);
+      localStorage.clear();
+      window.location.href = "/login"; // Handle any errors by redirecting to login
+    } finally {
+      setLoading(false); // Stop loading
     }
   };
 
+  useEffect(() => {
+    checkUserLoggedIn(); // Trigger check on mount
+  }, []);
+
   return (
-    <AuthContext.Provider value={{ user, setUser, checkUserLoggedIn }}>
+    <AuthContext.Provider value={{ user, setUser, checkUserLoggedIn, loading }}>
       {children}
     </AuthContext.Provider>
   );
